requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
PyAudio==0.2.11
pytz==2019.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
SpeechRecognition==3.8.1
sqlparse==0.3.0
Unipath==1.1
urllib3==1.24.2
